refactor(modules): split preprocessing into filter & imputer
