+++ src/urban_mapper/modules/enricher/aggregator/abc_aggregator.py
@@ -1,15 +1,14 @@
 from abc import ABC, abstractmethod
 import pandas as pd
 from beartype import beartype
-from osmnx_mapping.utils import require_arguments_not_none
+from urban_mapper.utils import require_arguments_not_none
 
 
+@beartype
 class BaseAggregator(ABC):
     @abstractmethod
-    @beartype
     def _aggregate(self, input_dataframe: pd.DataFrame) -> pd.Series: ...
 
-    @beartype
     @require_arguments_not_none(
         "input_dataframe", error_msg="No input dataframe provided.", check_empty=True
     )
