+++ src/urban_mapper/modules/enricher/aggregator/aggregators/count_aggregator.py
@@ -1,12 +1,12 @@
 from typing import Callable
 import pandas as pd
 from beartype import beartype
-from osmnx_mapping.modules.enricher.aggregator.abc_aggregator import BaseAggregator
-from osmnx_mapping.utils.helpers import require_attribute_columns
+from urban_mapper.modules.enricher.aggregator.abc_aggregator import BaseAggregator
+from urban_mapper.utils.helpers import require_attribute_columns
 
 
+@beartype
 class CountAggregator(BaseAggregator):
-    @beartype
     def __init__(
         self,
         group_by_column: str,
@@ -15,10 +15,10 @@ class CountAggregator(BaseAggregator):
         self.group_by_column = group_by_column
         self.count_function = count_function
 
-    @beartype
     @require_attribute_columns("input_dataframe", ["group_by_column"])
     def _aggregate(self, input_dataframe: pd.DataFrame) -> pd.Series:
         grouped = input_dataframe.groupby(self.group_by_column)
+
         return pd.Series(
             {
                 group_name: self.count_function(group_data)
