+++ src/urban_mapper/modules/enricher/aggregator/aggregators/simple_aggregator.py
@@ -1,11 +1,10 @@
 from typing import Callable
 import pandas as pd
 from beartype import beartype
-from osmnx_mapping.modules.enricher.aggregator.abc_aggregator import BaseAggregator
+from urban_mapper.modules.enricher.aggregator.abc_aggregator import BaseAggregator
 
 from typing import Dict
 
-from osmnx_mapping.utils import require_attribute_columns
 
 AGGREGATION_FUNCTIONS: Dict[str, Callable[[pd.Series], float]] = {
     "mean": pd.Series.mean,
@@ -16,8 +15,8 @@ AGGREGATION_FUNCTIONS: Dict[str, Callable[[pd.Series], float]] = {
 }
 
 
+@beartype
 class SimpleAggregator(BaseAggregator):
-    @beartype
     def __init__(
         self,
         group_by_column: str,
@@ -28,9 +27,7 @@ class SimpleAggregator(BaseAggregator):
         self.value_column = value_column
         self.aggregation_function = aggregation_function
 
-    @beartype
-    @require_attribute_columns("input_dataframe", ["group_by_column", "value_column"])
     def _aggregate(self, input_dataframe: pd.DataFrame) -> pd.Series:
-        return input_dataframe.groupby(self.group_by_column)[self.value_column].agg(
-            self.aggregation_function
-        )
+        grouped = input_dataframe.groupby(self.group_by_column)
+        aggregated = grouped[self.value_column].agg(self.aggregation_function)
+        return aggregated
