+++ src/urban_mapper/modules/enricher/factory/__init__.py
@@ -10,7 +10,6 @@ from .validation import (
     validate_group_by,
     validate_action,
     validate_aggregation_method,
-    validate_enricher_type,
 )
 
 from .preview import PreviewBuilder
@@ -24,6 +23,5 @@ __all__ = [
     "validate_group_by",
     "validate_action",
     "validate_aggregation_method",
-    "validate_enricher_type",
     "PreviewBuilder",
 ]
