+++ src/urban_mapper/modules/enricher/factory/validation.py
@@ -1,6 +1,5 @@
 from .config import EnricherConfig
-from .registries import ENRICHER_REGISTRY
-from osmnx_mapping.modules.enricher.aggregator.aggregators.simple_aggregator import (
+from urban_mapper.modules.enricher.aggregator.aggregators.simple_aggregator import (
     AGGREGATION_FUNCTIONS,
 )
 
@@ -19,11 +18,4 @@ def validate_aggregation_method(method: str):
     if method not in AGGREGATION_FUNCTIONS:
         raise ValueError(
             f"Unknown aggregation method '{method}'. Available: {list(AGGREGATION_FUNCTIONS.keys())}"
-        )
-
-
-def validate_enricher_type(enricher_type: str):
-    if enricher_type not in ENRICHER_REGISTRY:
-        raise ValueError(
-            f"Unknown enricher type '{enricher_type}'. Available: {list(ENRICHER_REGISTRY.keys())}"
         )
